% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWRstudioAddins.R
\name{insertRmdHeader}
\alias{insertRmdHeader}
\title{A Function for Generating R Markdown Header Labels}
\usage{
insertRmdHeader()
}
\arguments{
\item{NULL}{no arguments}
}
\description{
This function adds a basic Rmd header to cursor location
}
\examples{
insertRmdHeader
}
\keyword{Header}

