% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWRstudioAddins.R
\name{insertScriptHeader}
\alias{insertScriptHeader}
\title{A Function for Generating Script Header Labels}
\usage{
insertScriptHeader()
}
\arguments{
\item{NULL}{no arguments}
}
\description{
This function adds a header with title, date, and summary to cursor location
}
\examples{
insertScriptHeader
}
\keyword{Header}

